+++ src/main/java/com/example/JavaProject/controller/FavouritesController.java
@@ -1,7 +1,7 @@
 package com.example.JavaProject.controller;
 
 import com.example.JavaProject.response.RecipeResponse;
-import com.example.JavaProject.service.interfaces.RecipeService;
+import com.example.JavaProject.service.interfaces.FavouritesService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,23 +16,23 @@ import java.util.List;
 @RequiredArgsConstructor
 public class FavouritesController {
 
-    private final RecipeService recipeService;
+    private final FavouritesService favouritesService;
 
     @PutMapping("/recipes/{id}/favourites")
     @PreAuthorize("isAuthenticated()")
     public ResponseEntity<String> addRecipeToFavourite(@PathVariable long id) {
-        String message = recipeService.addToFavourite(id);
+        String message = favouritesService.addToFavourite(id);
         return new ResponseEntity<>(message, HttpStatus.OK);
     }
 
     @GetMapping("/recipes/favourites")
     public ResponseEntity<List<RecipeResponse>> getFavouriteRecipes() {
-        return new ResponseEntity<>(recipeService.getFavouriteRecipes(), HttpStatus.OK);
+        return new ResponseEntity<>(favouritesService.getFavouriteRecipes(), HttpStatus.OK);
     }
 
     @DeleteMapping("/recipes/{id}/favourites")
     public ResponseEntity<String> deleteFavouriteRecipe(@PathVariable Long id) {
-        String message = recipeService.deleteFavouriteRecipe(id);
+        String message = favouritesService.deleteFavouriteRecipe(id);
         return new ResponseEntity<>(message, HttpStatus.OK);
     }
 }
