+++ src/main/java/com/example/JavaProject/controller/LikesController.java
@@ -1,6 +1,7 @@
 package com.example.JavaProject.controller;
 
 import com.example.JavaProject.response.LikesCountResponse;
+import com.example.JavaProject.service.interfaces.LikesService;
 import com.example.JavaProject.service.interfaces.RecipeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -13,21 +14,21 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class LikesController {
 
-    private final RecipeService recipeService;
+    private final LikesService likesService;
 
     @GetMapping("/recipes/{id}/likes")
     public ResponseEntity<LikesCountResponse> getRecipeLikeCount(@PathVariable long id) {
-        return new ResponseEntity<>(recipeService.getRecipeLikes(id), HttpStatus.OK);
+        return new ResponseEntity<>(likesService.getRecipeLikes(id), HttpStatus.OK);
     }
 
     @PutMapping("/recipes/{id}/likes")
     public ResponseEntity<String> addRecipeLike(@PathVariable long id) {
-        return new ResponseEntity<>(recipeService.addLike(id), HttpStatus.OK);
+        return new ResponseEntity<>(likesService.addLike(id), HttpStatus.OK);
     }
 
     @PutMapping("recipes/{id}/unlikes")
     public ResponseEntity<String> removeRecipeLike(@PathVariable long id) {
-        return new ResponseEntity<>(recipeService.removeLike(id), HttpStatus.OK);
+        return new ResponseEntity<>(likesService.removeLike(id), HttpStatus.OK);
     }
 
 }
