+++ src/main/java/com/example/JavaProject/controller/RecipeController.java
@@ -7,7 +7,6 @@ import com.example.JavaProject.service.interfaces.RecipeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
