+++ src/main/java/com/example/JavaProject/service/implementation/RecipeServiceImpl.java
@@ -13,10 +13,8 @@ import com.example.JavaProject.repository.IngredientRepository;
 import com.example.JavaProject.repository.RecipeIngredientRepository;
 import com.example.JavaProject.repository.RecipeRepository;
 import com.example.JavaProject.repository.UserRepository;
-import com.example.JavaProject.response.LikesCountResponse;
 import com.example.JavaProject.response.RecipeResponse;
 import com.example.JavaProject.service.interfaces.AuthenticationService;
-import com.example.JavaProject.service.interfaces.EmailService;
 import com.example.JavaProject.service.interfaces.RecipeService;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -31,8 +29,6 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 @Service
 public class RecipeServiceImpl implements RecipeService {
-    private EmailService emailServiceImpl;
-
     private RecipeRepository recipeRepository;
     private IngredientRepository ingredientRepository;
     private RecipeIngredientRepository recipeIngredientRepository;
@@ -92,26 +88,6 @@ public class RecipeServiceImpl implements RecipeService {
         recipeResponse.setIngredients(ingredientDtos);
         return recipeResponse;
     }
-
-    @Override
-    public LikesCountResponse getRecipeLikes(long id) {
-        Optional<Recipe> recipe = recipeRepository.findById(id);
-
-        if (recipe.isEmpty()) {
-            throw new RuntimeException("Recipe doesn't exist");
-        }
-
-        Recipe foundRecipe = recipe.get();
-        if (isHidden(foundRecipe)) {
-            throw new ProfileHiddenException("Profile has been deleted");
-        }
-
-        int likesCount = (int) foundRecipe.getLikedby().stream().count();
-        return new LikesCountResponse(
-                likesCount
-        );
-    }
-
     @Override
     public String addRecipe(RecipeDto recipeDto) {
         Long userId = authenticationService.getCurrentUserId();
@@ -153,7 +129,6 @@ public class RecipeServiceImpl implements RecipeService {
 
         return "Added new recipe";
     }
-
 
     @Transactional
     @Override
@@ -231,158 +206,6 @@ public class RecipeServiceImpl implements RecipeService {
     }
 
     @Override
-    public String addLike(long id) {
-        Long userID = authenticationService.getCurrentUserId();
-
-        if (userID == null) {
-            return "User not logged in";
-        }
-
-        Optional<Recipe> foundRecipe = recipeRepository.findById(id);
-        if (foundRecipe.isEmpty()) {
-            return "Recipe not found";
-        }
-
-        Recipe recipe = foundRecipe.get();
-
-        boolean isAlreadyLiked = recipe.getLikedby().stream().anyMatch(user -> user.getId().equals(userID));
-        if (isAlreadyLiked) {
-            return "User already liked this recipe";
-        }
-
-        User user = userRepository.findById(userID).orElseThrow(() -> new RuntimeException("User Not Found"));
-
-        recipe.getLikedby().add(user);
-        recipeRepository.save(recipe);
-        String recipientEmail = recipe.getUser().getEmail();
-        System.out.println("Sending email to: " + recipientEmail);
-        emailServiceImpl.sendLikeNotification(recipe.getUser().getEmail());
-        return "Recipe liked successfully ";
-    }
-
-    public String removeLike(long id) {
-        Long userID = authenticationService.getCurrentUserId();
-
-        if (userID == null) {
-            return "User not logged in";
-        }
-
-        Optional<Recipe> foundRecipe = recipeRepository.findById(id);
-        if (foundRecipe.isEmpty()) {
-            return "Recipe not found";
-        }
-
-        Recipe recipe = foundRecipe.get();
-
-        boolean isLiked = recipe.getLikedby().stream().anyMatch(user -> user.getId().equals(userID));
-        if (!isLiked) {
-            return "User has not liked this recipe";
-        }
-
-        User user = userRepository.findById(userID).orElseThrow(() -> new RuntimeException("User Not Found"));
-        recipe.getLikedby().remove(user);
-        recipeRepository.save(recipe);
-
-        return "Recipe unliked successfully";
-    }
-
-
-    @Transactional
-    @Override
-    public String addToFavourite(long id) {
-        Long userID = authenticationService.getCurrentUserId();
-
-        if (userID == null) {
-            return "User not logged in";
-        }
-
-        Optional<Recipe> foundRecipe = recipeRepository.findById(id);
-        if (foundRecipe.isEmpty()) {
-            return "Recipe not found";
-        }
-
-        Recipe recipe = foundRecipe.get();
-
-        boolean isUserDeleted = recipe.getUser().isHidden();
-        if (isUserDeleted) {
-            return "User has been deleted";
-        }
-
-        boolean isAlreadyFavoured = recipe.getFavouritedBy().stream()
-                .anyMatch(user -> user.getId().equals(userID));
-        if (isAlreadyFavoured) {
-            return "User has already added this recipe to favourite";
-        }
-
-        User user = userRepository.findById(userID)
-                .orElseThrow(() -> new RuntimeException("User Not Found"));
-
-        recipe.getFavouritedBy().add(user);
-        recipeRepository.save(recipe);
-        return "Recipe added to favourites successfully ";
-    }
-
-    @Override
-    public List<RecipeResponse> getFavouriteRecipes() {
-        Long userID = authenticationService.getCurrentUserId();
-        Optional<User> updatedUser = userRepository.findById(userID);
-        if (updatedUser.isEmpty()) {
-            throw new RuntimeException("No user found with currently logged account.");
-        }
-
-        User user = updatedUser.get();
-
-        if (user.isHidden()) {
-            throw new ProfileHiddenException("Profile has been deleted");
-        }
-
-        List<RecipeResponse> recipeResponses = new ArrayList<>();
-        for (Recipe recipe : user.getFavourites()) {
-            if (recipe.getUser().isHidden()) {
-                continue;
-            }
-
-            RecipeDto recipeDto = recipeMapper.mapToDto(recipe);
-            RecipeResponse recipeResponse = new RecipeResponse();
-
-            List<IngredientDto> ingredientDtos = new ArrayList<>();
-            for (RecipeIngredient recipeIngredient : recipe.getRecipeIngredients()) {
-                ingredientDtos.add(ingredientsMapper.mapToDto(recipeIngredient));
-            }
-
-            recipeResponse.setIngredients(ingredientDtos);
-            recipeResponse.setRecipeName(recipeDto.getRecipeName());
-            recipeResponse.setPreparationTime(recipeDto.getPreparationTime());
-            recipeResponse.setDifficulty(recipeDto.getDifficulty());
-            recipeResponse.setAuthor(recipeDto.getAuthor());
-            recipeResponse.setLikes(recipeDto.getLikes());
-            recipeResponses.add(recipeResponse);
-        }
-
-        return recipeResponses;
-    }
-
-    @Override
-    public String deleteFavouriteRecipe(Long id) {
-        Long userID = authenticationService.getCurrentUserId();
-        User user = userRepository.findById(userID)
-                .orElseThrow(() -> new RuntimeException("No user found with currently logged account."));
-
-        Recipe recipe = recipeRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Recipe not found"));
-
-        if (!user.getFavourites().contains(recipe)) {
-            return "Recipe is not in favourites.";
-        }
-
-        user.getFavourites().remove(recipe);
-
-        userRepository.save(user);
-
-        return "Recipe removed from favourites successfully";
-    }
-
-    @Override
     public List<RecipeResponse> findRecipesByIngredient(String ingredientName) {
         List<Recipe> recipes = recipeRepository.findRecipesByIngredientName(ingredientName);
 
@@ -414,6 +237,4 @@ public class RecipeServiceImpl implements RecipeService {
                 .collect(Collectors.toList());
         return recipeDtos;
     }
-
-
 }
